Update pocketmine.yml
Hoping to have this default to enabled with the option to disable. We use this file when users wish to see a Web console. - We can do it without the change but would no longer be able to use this embedded pocketmine.yml file and would apply our own when users update.  Thoughts?