Joining fix
When devices with less or 512 MB RAM join, they see "Building Terrain" screen and immediatly MCPE crashes. Only way how to fix that is set chunkRadius to 10 or less. (even if $server->chunkRadius is 10) it doesn't work and MCPE crashes.

Note: It should be better to have option if it should be 10 (512 MB can join) or 30 (only 1024 MB and higher can join)